Refactor router in a seperate router.go file
