+++ go.mod
@@ -4,7 +4,6 @@ go 1.16
 
 require (
 	github.com/go-chi/chi v1.5.4
-	github.com/go-chi/chi/v5 v5.0.1
 	github.com/mitchellh/go-homedir v1.1.0
 	github.com/spf13/cobra v1.1.3
 	github.com/spf13/viper v1.7.1
